Use context for web api
